+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fixed import of `delaunay_from_points` in `Mesh.from_points`.
+- More control over drawing of text labels in Rhino.
 
 ### Removed